# go-server/internal/gen/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stick-It RESP API
  version: 1.0.1
servers:
- url: https://stick-it.lr-projects.de
- url: http://localhost:8081
...
paths:
  /api/v2/groups:
    get:
      operationId: getGroupsByIds
      parameters:
      - description: Comma-separated list of group IDs
        in: query
        name: ids
        schema:
          items:
            $ref: "#/components/schemas/id"
          type: array
      - description: search term used to find a group name
        in: query
        name: search
        schema:
          $ref: "#/components/schemas/shortString"
      - description: this is the userId used when withUser is set to true or false
        in: query
        name: userId
        schema:
          $ref: "#/components/schemas/id"
      - description: use false if user groups should be included and true if user
          groups should be excluded from search
        in: query
        name: withUser
        schema:
          type: boolean
      - description: use false if profile picture should not be returned and true
          if it should. Defaults to false
        in: query
        name: withImages
        schema:
          default: false
          type: boolean
      - description: page number
        in: query
        name: page
        schema:
          type: integer
      - description: page size. Defaults to 20
        in: query
        name: size
        schema:
          default: 20
          type: integer
      - description: only include groups that have been updated after this date. If
          set all deleted groups after this time are returned.
        in: query
        name: updatedAfter
        schema:
          $ref: "#/components/schemas/date"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupsSyncDto"
          description: Successful response
      security:
      - token: []
      summary: Get groups by IDs
      tags:
      - groups
    post:
      operationId: addGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createGroupDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupDto"
          description: Successful response
        "400":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Image can not be used or is not square
      security:
      - token: []
      summary: Create a new group
      tags:
      - groups
  /api/v2/groups/{groupId}:
    delete:
      operationId: deleteGroup
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          description: Deleted
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Delete a group by ID
      tags:
      - groups
    get:
      operationId: getGroup
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupDto"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get a group by ID
      tags:
      - groups
    put:
      operationId: updateGroup
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateGroupDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupDto"
          description: Successful response
        "400":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Image can not be used or is not square
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Update a group by ID
      tags:
      - groups
  /api/v2/groups/{groupId}/profile_image:
    get:
      operationId: getGroupProfileImage
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get profile of group
      tags:
      - groups
  /api/v2/groups/{groupId}/profile_image_small:
    get:
      operationId: getGroupProfileImageSmall
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get small profile image url of group
      tags:
      - groups
  /api/v2/groups/{groupId}/pin_image:
    get:
      operationId: getGroupPinImage
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/png:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get pin image of group
      tags:
      - groups
  /api/v2/groups/{groupId}/description:
    get:
      operationId: getGroupDescription
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/shortString"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get description of group
      tags:
      - groups
  /api/v2/groups/{groupId}/link:
    get:
      operationId: getGroupLink
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/shortString"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get link of group
      tags:
      - groups
  /api/v2/groups/{groupId}/admin:
    get:
      operationId: getGroupAdmin
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/shortString"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get admin of group
      tags:
      - groups
  /api/v2/groups/{groupId}/invite_url:
    get:
      operationId: getGroupInviteUrl
      parameters:
      - description: group id
        in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/shortString"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get invite url of group
      tags:
      - groups
...
components:
  schemas:
    id:
      format: uuid
      type: string
    shortString:
      maxLength: 256
      type: string
    date:
      example: 2017-01-10T14:55:32+01:00
      format: date-time
      type: string
    groupsSyncDto:
      example:
        deleted:
        - null
        - null
        items:
        - lastUpdated: 2017-01-10T14:55:32+01:00
          bestSeason:
            season:
              month: 6
              year: 1
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              seasonNumber: 5
            rank: 2
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            points: 5
          profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
          visibility: 0
          group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          invite_url: invite_url
          name: name
          link: link
          pinImage: https://minio.lr-projects/stick-it/1/1.png
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          profileImage: https://minio.lr-projects/stick-it/1/1.png
        - lastUpdated: 2017-01-10T14:55:32+01:00
          bestSeason:
            season:
              month: 6
              year: 1
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              seasonNumber: 5
            rank: 2
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            points: 5
          profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
          visibility: 0
          group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          invite_url: invite_url
          name: name
          link: link
          pinImage: https://minio.lr-projects/stick-it/1/1.png
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          profileImage: https://minio.lr-projects/stick-it/1/1.png
      properties:
        items:
          items:
            $ref: "#/components/schemas/groupDto"
          type: array
        deleted:
          items:
            $ref: "#/components/schemas/id"
          type: array
      required:
      - deleted
      - items
      type: object
    createGroupDto:
      properties:
        description:
          maxLength: 256
          type: string
        name:
          maxLength: 256
          type: string
        profileImage:
          example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=
          format: byte
          type: string
        visibility:
          description: "The visibility of the group. 0 for public, 1 for private"
          type: integer
        groupAdmin:
          format: uuid
          type: string
        link:
          maxLength: 256
          type: string
        userId:
          format: uuid
          type: string
      required:
      - description
      - groupAdmin
      - name
      - profileImage
      - visibility
      type: object
    groupDto:
      example:
        lastUpdated: 2017-01-10T14:55:32+01:00
        bestSeason:
          season:
            month: 6
            year: 1
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            seasonNumber: 5
          rank: 2
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          points: 5
        profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
        visibility: 0
        group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        invite_url: invite_url
        name: name
        link: link
        pinImage: https://minio.lr-projects/stick-it/1/1.png
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        profileImage: https://minio.lr-projects/stick-it/1/1.png
      properties:
        id:
          format: uuid
          type: string
        description:
          maxLength: 256
          type: string
        invite_url:
          maxLength: 256
          type: string
        name:
          maxLength: 256
          type: string
        visibility:
          description: "The visibility of the group. 0 for public, 1 for private"
          type: integer
        group_admin:
          format: uuid
          type: string
        link:
          maxLength: 256
          type: string
        lastUpdated:
          example: 2017-01-10T14:55:32+01:00
          format: date-time
          type: string
        profileImage:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        profileImageSmall:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        pinImage:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        bestSeason:
          $ref: "#/components/schemas/seasonItemDto"
      required:
      - id
      - name
      - visibility
      type: object
    longString:
      maxLength: 500
      type: string
    updateGroupDto:
      properties:
        description:
          maxLength: 256
          type: string
        name:
          maxLength: 256
          type: string
        profileImage:
          example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=
          format: byte
          type: string
        visibility:
          description: "The visibility of the group. 0 for public, 1 for private"
          type: integer
        groupAdmin:
          format: uuid
          type: string
        link:
          maxLength: 256
          type: string
      type: object
    imageUrl:
      example: https://minio.lr-projects/stick-it/1/1.png
      format: url
      type: string
...
    seasonItemDto:
      example:
        season:
          month: 6
          year: 1
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          seasonNumber: 5
        rank: 2
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        points: 5
      properties:
        id:
          format: uuid
          type: string
        season:
          $ref: "#/components/schemas/seasonDto"
        points:
          description: Number of sticks
          type: integer
        rank:
          description: Rank reached at the end of the season
          type: integer
      required:
      - id
      - points
      - rank
      - season
      type: object
...
    seasonDto:
      example:
        month: 6
        year: 1
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        seasonNumber: 5
      properties:
        id:
          format: uuid
          type: string
        month:
          type: integer
        year:
          type: integer
        seasonNumber:
          type: integer
      required:
      - id
      - month
      - seasonNumber
      - year
      type: object
...
  securitySchemes:
    token:
      bearerFormat: JWT
      scheme: bearer
      type: http