# go-server/internal/gen/server/api_groups.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Stick-It RESP API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.1
 */

package genserver

import (
	"encoding/json"
	"net/http"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
)

// GroupsAPIController binds http requests to an api service and writes the service results to the http response
type GroupsAPIController struct {
	service      GroupsAPIServicer
	errorHandler ErrorHandler
}

// GroupsAPIOption for how the controller is set up.
type GroupsAPIOption func(*GroupsAPIController)

// WithGroupsAPIErrorHandler inject ErrorHandler into controller
func WithGroupsAPIErrorHandler(h ErrorHandler) GroupsAPIOption {
	return func(c *GroupsAPIController) {
		c.errorHandler = h
	}
}

// NewGroupsAPIController creates a default api controller
func NewGroupsAPIController(s GroupsAPIServicer, opts ...GroupsAPIOption) *GroupsAPIController {
	controller := &GroupsAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the GroupsAPIController
func (c *GroupsAPIController) Routes() Routes {
	return Routes{
		"GetGroupsByIds": Route{
			"GetGroupsByIds",
			strings.ToUpper("Get"),
			"/api/v2/groups",
			c.GetGroupsByIds,
		},
		"AddGroup": Route{
			"AddGroup",
			strings.ToUpper("Post"),
			"/api/v2/groups",
			c.AddGroup,
		},
		"GetGroup": Route{
			"GetGroup",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}",
			c.GetGroup,
		},
		"UpdateGroup": Route{
			"UpdateGroup",
			strings.ToUpper("Put"),
			"/api/v2/groups/{groupId}",
			c.UpdateGroup,
		},
		"DeleteGroup": Route{
			"DeleteGroup",
			strings.ToUpper("Delete"),
			"/api/v2/groups/{groupId}",
			c.DeleteGroup,
		},
		"GetGroupProfileImage": Route{
			"GetGroupProfileImage",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/profile_image",
			c.GetGroupProfileImage,
		},
		"GetGroupProfileImageSmall": Route{
			"GetGroupProfileImageSmall",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/profile_image_small",
			c.GetGroupProfileImageSmall,
		},
		"GetGroupPinImage": Route{
			"GetGroupPinImage",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/pin_image",
			c.GetGroupPinImage,
		},
		"GetGroupDescription": Route{
			"GetGroupDescription",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/description",
			c.GetGroupDescription,
		},
		"GetGroupLink": Route{
			"GetGroupLink",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/link",
			c.GetGroupLink,
		},
		"GetGroupAdmin": Route{
			"GetGroupAdmin",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/admin",
			c.GetGroupAdmin,
		},
		"GetGroupInviteUrl": Route{
			"GetGroupInviteUrl",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/invite_url",
			c.GetGroupInviteUrl,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the GroupsAPIController
func (c *GroupsAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetGroupsByIds",
			strings.ToUpper("Get"),
			"/api/v2/groups",
			c.GetGroupsByIds,
		},
		Route{
			"AddGroup",
			strings.ToUpper("Post"),
			"/api/v2/groups",
			c.AddGroup,
		},
		Route{
			"GetGroup",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}",
			c.GetGroup,
		},
		Route{
			"UpdateGroup",
			strings.ToUpper("Put"),
			"/api/v2/groups/{groupId}",
			c.UpdateGroup,
		},
		Route{
			"DeleteGroup",
			strings.ToUpper("Delete"),
			"/api/v2/groups/{groupId}",
			c.DeleteGroup,
		},
		Route{
			"GetGroupProfileImage",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/profile_image",
			c.GetGroupProfileImage,
		},
		Route{
			"GetGroupProfileImageSmall",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/profile_image_small",
			c.GetGroupProfileImageSmall,
		},
		Route{
			"GetGroupPinImage",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/pin_image",
			c.GetGroupPinImage,
		},
		Route{
			"GetGroupDescription",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/description",
			c.GetGroupDescription,
		},
		Route{
			"GetGroupLink",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/link",
			c.GetGroupLink,
		},
		Route{
			"GetGroupAdmin",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/admin",
			c.GetGroupAdmin,
		},
		Route{
			"GetGroupInviteUrl",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/invite_url",
			c.GetGroupInviteUrl,
		},
	}
}

// GetGroupsByIds - Get groups by IDs
func (c *GroupsAPIController) GetGroupsByIds(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	var idsParam []string
	if query.Has("ids") {
		idsParam = strings.Split(query.Get("ids"), ",")
	}
	var searchParam string
	if query.Has("search") {
		param := query.Get("search")

		searchParam = param
	} else {
	}
	var userIdParam string
	if query.Has("userId") {
		param := query.Get("userId")

		userIdParam = param
	} else {
	}
	var withUserParam bool
	if query.Has("withUser") {
		param, err := parseBoolParameter(
			query.Get("withUser"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "withUser", Err: err}, nil)
			return
		}

		withUserParam = param
	} else {
	}
	var withImagesParam bool
	if query.Has("withImages") {
		param, err := parseBoolParameter(
			query.Get("withImages"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "withImages", Err: err}, nil)
			return
		}

		withImagesParam = param
	} else {
		var param bool = false
		withImagesParam = param
	}
	var pageParam int32
	if query.Has("page") {
		param, err := parseNumericParameter[int32](
			query.Get("page"),
			WithParse[int32](parseInt32),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "page", Err: err}, nil)
			return
		}

		pageParam = param
	} else {
	}
	var sizeParam int32
	if query.Has("size") {
		param, err := parseNumericParameter[int32](
			query.Get("size"),
			WithParse[int32](parseInt32),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "size", Err: err}, nil)
			return
		}

		sizeParam = param
	} else {
		var param int32 = 20
		sizeParam = param
	}
	var updatedAfterParam time.Time
	if query.Has("updatedAfter") {
		param, err := parseTime(query.Get("updatedAfter"))
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "updatedAfter", Err: err}, nil)
			return
		}

		updatedAfterParam = param
	} else {
	}
	result, err := c.service.GetGroupsByIds(r.Context(), idsParam, searchParam, userIdParam, withUserParam, withImagesParam, pageParam, sizeParam, updatedAfterParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// AddGroup - Create a new group
func (c *GroupsAPIController) AddGroup(w http.ResponseWriter, r *http.Request) {
	var createGroupDtoParam CreateGroupDto
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&createGroupDtoParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertCreateGroupDtoRequired(createGroupDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertCreateGroupDtoConstraints(createGroupDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.AddGroup(r.Context(), createGroupDtoParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroup - Get a group by ID
func (c *GroupsAPIController) GetGroup(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroup(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateGroup - Update a group by ID
func (c *GroupsAPIController) UpdateGroup(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var updateGroupDtoParam UpdateGroupDto
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&updateGroupDtoParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertUpdateGroupDtoRequired(updateGroupDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertUpdateGroupDtoConstraints(updateGroupDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateGroup(r.Context(), groupIdParam, updateGroupDtoParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteGroup - Delete a group by ID
func (c *GroupsAPIController) DeleteGroup(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.DeleteGroup(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupProfileImage - Get profile of group
func (c *GroupsAPIController) GetGroupProfileImage(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetGroupProfileImage(r.Context(), groupIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupProfileImageSmall - Get small profile image url of group
func (c *GroupsAPIController) GetGroupProfileImageSmall(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetGroupProfileImageSmall(r.Context(), groupIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupPinImage - Get pin image of group
func (c *GroupsAPIController) GetGroupPinImage(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetGroupPinImage(r.Context(), groupIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupDescription - Get description of group
func (c *GroupsAPIController) GetGroupDescription(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroupDescription(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupLink - Get link of group
func (c *GroupsAPIController) GetGroupLink(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroupLink(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupAdmin - Get admin of group
func (c *GroupsAPIController) GetGroupAdmin(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroupAdmin(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetGroupInviteUrl - Get invite url of group
func (c *GroupsAPIController) GetGroupInviteUrl(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroupInviteUrl(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}